Backend/requirements.txt
flask
flask-sqlalchemy 
flask-cors
psycopg2-binary
flask_jwt_extended
flask-cors

pytest
pytest-flask>=1.3
python-dotenv
